# services/user/routes.go
package user

import (
	"fmt"
	"net/http"

	"github.com/go-playground/validator/v10"
	"github.com/thangsuperman/bee-happy/config"
	"github.com/thangsuperman/bee-happy/services/auth"
	"github.com/thangsuperman/bee-happy/types"
	"github.com/thangsuperman/bee-happy/utils"

	"github.com/gorilla/mux"
)

type Handler struct {
	store types.UserStore
}

func NewHandler(store types.UserStore) *Handler {
	return &Handler{store: store}
}

func (h *Handler) RegisterRoutes(router *mux.Router) {
	router.HandleFunc("/login", h.handleLogin).Methods("POST")
	router.HandleFunc("/register", h.handleRegister).Methods("POST")
}

// handleLogin   Login
// @Summary		   Login
// @Description  Login by email and password
// @Tags			   Auth
// @Accept			 json
// @Produce		   json
// @Param        request body types.LoginUserPayload true "Payload of login user"
// @Success		   200 {object} types.BaseResponse "Token"
// @Failure		   400 {object} types.ErrorLoginResponse  "Invalid payload [errors]"
// @Router			 /api/v1/login [post]
func (h *Handler) handleLogin(w http.ResponseWriter, r *http.Request) {
	var payload types.LoginUserPayload
	if err := utils.ParseJSON(r, &payload); err != nil {
		utils.WriteError(w, http.StatusBadRequest, err)
		return
	}

	if err := utils.Validate.Struct(payload); err != nil {
		errors := err.(validator.ValidationErrors)
		utils.WriteError(w, http.StatusBadRequest, fmt.Errorf("invalid payload %w", errors))
		return
	}

	u, err := h.store.GetUserByEmail(payload.Email)
	if err != nil {
		utils.WriteError(w, http.StatusBadRequest, fmt.Errorf("not found, invalid email or password"))
		return
	}

	if !auth.ComparePassword(u.Password, []byte(payload.Password)) {
		utils.WriteError(w, http.StatusBadRequest, fmt.Errorf("password does not correct, please retry!"))
		return
	}

	secret := []byte(config.Envs.JWTSecret)
	token, err := auth.CreateJWT(secret, u.ID)
	if err != nil {
		utils.WriteError(w, http.StatusInternalServerError, err)
		return
	}

	utils.WriteJSON(w, http.StatusOK, types.BaseResponse{
		Message: "Login successfully",
		Metadata: map[string]any{
			"token":     token,
			"user_id":   u.ID,
			"full_name": fmt.Sprintf("%s %s", u.FirstName, u.LastName),
			"email":     u.Email,
		},
	})
}

// handleLogin   Register
// @Summary		   Register a new account
// @Description  Register with credentials
// @Tags			   Auth
// @Accept			 json
// @Produce		   json
// @Param        request body types.RegisterUserPayload true "Payload of register user account"
// @Failure		   400 {object} types.ErrorEmailAlreadyExists "Email already exists"
// @Router			 /api/v1/register [post]
func (h *Handler) handleRegister(w http.ResponseWriter, r *http.Request) {
	var payload types.RegisterUserPayload
	if err := utils.ParseJSON(r, &payload); err != nil {
		utils.WriteError(w, http.StatusBadRequest, err)
		return
	}

	if err := utils.Validate.Struct(payload); err != nil {
		errors := err.(validator.ValidationErrors)
		utils.WriteError(w, http.StatusBadRequest, fmt.Errorf("invalid payload %w", errors))
		return
	}

	_, err := h.store.GetUserByEmail(payload.Email)
	if err == nil {
		utils.WriteError(w, http.StatusBadRequest, fmt.Errorf("user with email %s already exists", payload.Email))
		return
	}

	hashedPassword, err := auth.HashPassword(payload.Password)
	if err != nil {
		utils.WriteError(w, http.StatusInternalServerError, err)
		return
	}

	err = h.store.CreateUser(types.User{
		FirstName:   payload.FirstName,
		LastName:    payload.LastName,
		Email:       payload.Email,
		DateOfBirth: payload.DateOfBirth,
		Password:    hashedPassword,
	})
	if err != nil {
		utils.WriteError(w, http.StatusInternalServerError, err)
	}

	utils.WriteJSON(w, http.StatusCreated, types.BaseResponse{
		Message:  "Register successfully",
		Metadata: nil,
	})
}